rename SCIPION_LAUNCH_CMD with SCIPION_TEST_CMD Config class moved to it's own module schedule: removed tk ghost, due to errors and very likely not needed any more.
